requirements.txt
typer[all] >= 0.7.0
pandas >= 1.5.1
boto3 >= 1.24.73
numpy >= 1.23.5
pyarrow >= 10.0.1
loguru
python-decouple